rpc临时文件支持持久化

+++ src/rpc/File.php
@@ -6,15 +6,27 @@ use Throwable;
 
 class File extends \think\File
 {
+    protected $persistent = false;
+
+    /**
+     * 持久化该文件
+     */
+    public function persists()
+    {
+        $this->persistent = true;
+    }
+
     public function __destruct()
     {
         //销毁时删除临时文件
-        try {
-            if (file_exists($this->getPathname())) {
-                unlink($this->getPathname());
-            }
-        } catch (Throwable $e) {
+        if ($this->persistent) {
+            try {
+                if (file_exists($this->getPathname())) {
+                    unlink($this->getPathname());
+                }
+            } catch (Throwable $e) {
 
+            }
         }
     }
 }
